_metaData is now 64-bit... A bit more printing when transmitting payloads...

+++ Core/class_register.h
@@ -57,7 +57,7 @@ namespace	mBrane{
 		public:
 			char		class_name[255];
 			static	const	uint16	NoClass;	//	max: 0xFFFE classes
-			template<class	C,class	M>	static	uint32	Load();	//	returns class meta data (Cf payload.h)
+			template<class	C,class	M>	static	uint64	Load();	//	returns class meta data (Cf payload.h)
 			static	uint16	Load(const	char	*className);
 			static	ClassRegister	*Get(uint16	CID);
 			static	uint16	GetCID(const	char	*className);
