+++ Core/payload.h
@@ -92,7 +92,7 @@ namespace	mBrane{
 		protected:
 			int64	_node_recv_ts;	//	not transmitted
 			int64	_recv_ts;		//	not transmitted
-			uint32	_metaData;		//	[cid(16)|reserved(12)|category(2)|allocation scheme(2)]
+			uint64	_metaData;		//	[cid(16)|reserved(12)|category(2)|allocation scheme(2)]
 			int64	_node_send_ts;
 			int64	_send_ts;
 			_Payload();
@@ -111,7 +111,7 @@ namespace	mBrane{
 		class	dll	_RPayload:
 		public	__Payload{
 		protected:
-			uint32	_metaData;	//	[cid(16)|reserved(14)|allocation scheme(2)]
+			uint64	_metaData;	//	[cid(16)|reserved(14)|allocation scheme(2)]
 			_RPayload();
 		public:
 			virtual	~_RPayload();
@@ -174,7 +174,7 @@ namespace	mBrane{
 		protected:
 			___Payload();
 		public:
-			static	const	uint32	_MetaData;
+			static	const	uint64	_MetaData;
 			static	void			*New(uint32	size);	//	to initialize the _vftable on recv(); size used for non-standard cases (like Storage<T>), i.e. when the actual size is not sizeof the class
 			static	uint16			CID();
 			static	size_t			Offset();	//	to metadata from this
